+++ apps/readest-app/src/app/reader/components/notebook/Notebook.tsx
@@ -21,7 +21,7 @@ import { saveSysSettings } from '@/helpers/settings';
 import { NOTE_PREFIX } from '@/types/view';
 import useShortcuts from '@/hooks/useShortcuts';
 import BooknoteItem from '../sidebar/BooknoteItem';
-import AIAssistant from '../sidebar/AIAssistant';
+import AIAssistant from './AIAssistant';
 import NotebookHeader from './Header';
 import NoteEditor from './NoteEditor';
 import SearchBar from './SearchBar';
