+++ apps/readest-app/src/components/assistant/TooltipIconButton.tsx
@@ -4,7 +4,7 @@ import { forwardRef, type ComponentPropsWithRef } from 'react';
 import { Slottable } from '@radix-ui/react-slot';
 import { Tooltip, TooltipContent, TooltipTrigger, TooltipProvider } from '@/components/ui/tooltip';
 import { Button } from '@/components/ui/button';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/tailwind';
 
 export type TooltipIconButtonProps = ComponentPropsWithRef<typeof Button> & {
   tooltip: string;
