+++ apps/readest-app/src/components/primitives/button-group.tsx
@@ -1,7 +1,7 @@
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/tailwind';
 import { Separator } from '@/components/ui/separator';
 
 const buttonGroupVariants = cva(
