+++ apps/readest-app/src/components/primitives/dialog.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 import { X } from 'lucide-react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/tailwind';
 
 const Dialog = DialogPrimitive.Root;
